feat: migrate to TypeScript and enhance loading bar options
