+++ vite.config.js
@@ -4,7 +4,7 @@ import { resolve } from 'path'
 export default defineConfig({
   build: {
     lib: {
-      entry: resolve(__dirname, 'src/index.js'),
+      entry: resolve(__dirname, 'src/index.ts'),
       name: 'animate-loading',
       fileName: (format) => `main.js`,
       formats: ['umd']
